trying branch

+++ 01- Introductory Problems/06-NumberSpiral.cpp
@@ -48,7 +48,7 @@ int main()
 	freopen("input.txt", "r", stdin);
 	freopen("output.txt", "w", stdout);
 	#endif
-	int t; cin >> t;
+	ll t; cin >> t;
     while (t--)
     {
         ll y, x; cin >> y >> x;
